requirements.txt
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-speech==1.41.1
azure-common==1.1.28
azure-core==1.31.0
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
isodate==0.7.0
lxml==5.3.0
pydub==0.25.1
pypandoc==1.14
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
